Issue 323, add work functionality in collection
