+++ app/controllers/dashboard_controller.rb
@@ -1,5 +1,7 @@
 class DashboardController < ApplicationController
 
+  include AddWorkHelper
+
   before_filter :authorized?, :only => [:owner, :staging, :omeka, :startproject]
   before_filter :get_data, :only => [:owner, :staging, :omeka, :upload, :new_upload, :startproject]
 
@@ -39,6 +41,7 @@ class DashboardController < ApplicationController
   end
 
   # Owner Dashboard - start project
+  #other methods in AddWorkHelper
   def startproject
     @document_upload = DocumentUpload.new
     @document_upload.collection=@collection
@@ -51,40 +54,6 @@ class DashboardController < ApplicationController
   # Owner Dashboard - list of works
   def owner
   end
-
-  # Owner Dashboard - staging area
-  def staging
-  end
-
-  # Owner Dashboard - omeka import
-  def omeka
-    @omeka_items = OmekaItem.all
-    @omeka_sites = current_user.omeka_sites
-  end
-
-  # Owner Dashboard - upload document
-  def upload
-    @document_upload = DocumentUpload.new
-  end
-
-  def new_upload
-    @document_upload = DocumentUpload.new(params[:document_upload])
-    @document_upload.user = current_user
-
-    if @document_upload.save
-        if SMTP_ENABLED
-          flash[:notice] = "Document has been uploaded and will be processed shortly. We'll email you at #{@document_upload.user.email} when ready."
-          SystemMailer.new_upload(@document_upload).deliver!
-        else
-          flash[:notice] = "Document has been uploaded and will be processed shortly. Reload this page in a few minutes to see it."
-        end
-      @document_upload.submit_process
-      ajax_redirect_to controller: 'collection', action: 'show', collection_id: @document_upload.collection.id
-    else
-      render action: 'upload'
-    end
-  end
-  
 
   # Editor Dashboard - watchlist
   def watchlist
