+++ app/views/collection/new_work.html.slim
@@ -1,0 +1,3 @@
+=render({ :partial => '/shared/collection_tabs', :locals => { :selected => 4, :collection_id => @collection.id } })
+
+=render({ :partial => '/dashboard/new_work', :collection_id => @collection.id })
