+++ app/views/collection/show.html.slim
@@ -25,9 +25,14 @@
               span(style="width:#{progress_annotated}%")
               span(style="width:#{progress_transcribed - progress_annotated}%")
 
+    p.nodata_text = link_to 'Add a new work', { :controller => 'collection', :action => 'new_work', :collection_id => @collection.id }
+    
+
     -unless @collection.categories.empty?
       h3 Subject Categories
       p.big ==@collection.categories.where('parent_id is null').map { |c| link_to c.title, { :controller => 'article', :action => 'list', :collection_id => @collection.id, anchor: "category-#{c.id}" } }.join(', ')
+
+    
 
     .collection-users.justified
       .justified_box>
