+++ app/views/dashboard/startproject.html.slim
@@ -1,69 +1,3 @@
 =render({ :partial => 'owner_header', :locals => { :selected => 5 } })
 
-.columns
-  article.maincol
-    h3 Upload PDF or ZIP File
-    p.big Using this form you can import your page images into FromThePage. You should select a target collection to import into and attach a ZIP or PDF file containing page images to be used for your project.
-    =form_for :document_upload, url: { action: 'new_upload' } do |f|
-      =validation_summary @document_upload.errors
-      table.form
-        tr
-          td =f.select :collection_id, options_from_collection_for_select(current_user.collections, 'id', 'title', @document_upload.collection_id), prompt: '- Select a collection -'
-        tr
-          td
-            .input-file
-              =f.file_field :file, tabindex: '-1', accept: 'application/pdf,application/zip,application/x-zip,application/x-zip-compressed'
-              input(type="text" tabindex="-1" placeholder="Click to browse a file..." readonly)
-              button(type="button") Browse
-      ul
-        li ZIP files may contain folders containing images, PDFs, or folders containing pdfs.
-        li Each folder will be treated as a different document, so do not mix pages from different documents in the same folder.
-        li Each PDF will be treated as its own document, so do not split pages from the same document among more than one PDF.
-        li For example a ZIP file with 3 images, 2 PDFs, and 1 folder containing 5 more images would create 4 works: the top level images in one, each PDF in their own work, and a last work containing the 5 images from the folder.
-      p.aright
-        =f.button 'Upload File', class: 'big'
-    
-    h3 Import from an IIIF Repository
-    p.fgfaded Shared Canvas Universe
-    ul.omeka-sites
-      -@universe_collections.each do |sc_collection|
-        li.omeka-site
-          h5.omeka-site_title =link_to sc_collection.label, {:controller => 'sc_collections', :action => 'explore', :at_id => CGI::escape(sc_collection.at_id)}    
-      p.fgfaded Previously Viewed Collections
-      ul.omeka-sites
-        -@sc_collections.each do |sc_collection|
-          li.omeka-site
-            h5.omeka-site_title =link_to sc_collection.label, {:controller => 'sc_collections', :action => 'explore', :at_id => CGI::escape(sc_collection.at_id)}
-    =form_tag({ controller: 'sc_collections', action: 'explore_manifest' }) do |f|
-      table.form
-          td
-            .input-file
-              =text_field_tag :at_id
-              =button_tag 'Import'
-    
-    h3 Import from the Internet Archive
-    =link_to 'Import From Archive.org', { :controller => 'ia', :action => 'ia_book_form' }, :data => { litebox: { hash: 'import-ia' } }
-    h3 Import from Omeka
-    -if @omeka_sites.blank?
-      p.fgfaded You didn't connect any Omeka sites. Use the button below to create a new Omeka connection.
-    -else
-      p.fgfaded This is the list of connected Omeka sites. Click one to explore collections and import items.
-      ul.omeka-sites
-        -@omeka_sites.each do |omeka_site|
-          li.omeka-site
-            h5.omeka-site_title =link_to omeka_site.title, omeka_site
-            small.omeka-site_url =omeka_site.api_url
-            dl.dropdown.right
-              dt =svg_symbol '#icon-list', class: 'icon'
-              dd
-                =link_to 'Edit Site', edit_omeka_site_path(omeka_site), :data => { litebox: { hash: "edit-connection-#{omeka_site.id}" } }
-                hr
-                =link_to 'Delete Site', omeka_site, method: :delete, data: { confirm: 'Are you sure? You will need to re-create the site confiuration to import from this site again!' }, class: 'fgred'
-    =link_to 'Add Source', { :controller => 'omeka_sites', :action => 'new' }, :data => { litebox: { hash: 'create-connection' } }, class: 'button outline'
-  aside.sidecol
-    h4 Page Image Guidelines
-    .fglight
-      p PNG, GIF, and JPG files are all acceptible.
-      p Images should be oriented so that they are right-side-up.
-      p Images should be split down the spine, so that only one page is visible on an image.<br> (See the <a href="https://github.com/benwbrum/autosplit">autosplit</a> tool if you need to split two-page images.)
-      p Images should be named so that an alphabetical sort will result in the correct page order.<br> (This may require "zero-padding" for any page numbers: <code>page_09.jpg, page_10.jpg</code> will sort correctly, but <code>page_9.jpg, page_10.jpg</code> will not.)
+=render 'new_work'
